Reorganize application state
